+++ Source/Components/GAMs/SimulinkWrapperGAM/SimulinkClasses.h
@@ -216,7 +216,7 @@ public:
     /**
      * @brief Print information about the signal.
      */
-    void PrintSignal(const uint32 maxNameLength = 0u);
+    void PrintSignal(const uint64 maxNameLength = 0u);
 };
 
 /*---------------------------------------------------------------------------*/
@@ -291,7 +291,7 @@ public:
     /**
      * @brief Print information about this port.
      */
-    void PrintPort(const uint32 maxNameLength);
+    void PrintPort(const uint64 maxNameLength);
     
     /**
      * @brief Copy data from the associated MARTe2 signal to the associated model port.
